Make the window widget inform it's listeners when the close() method is called on it.
Refactor the ColumnModel and ColumnConfiguration so that they are interfaces with default implementations - changes include modifications to how they are converted from JSON to compensate.

Add the capability to listen to the GridPanel for events generated by the client for removing a column and requesting to add a column.

Updated Echo3 jars as the support for complex properties is required in order for changes to the ColumnModel to be sent to the client.